Time: 41 ms (28.19%), Space: 46.8 MB (76.68%) - LeetHub

+++ 1283-find-the-smallest-divisor-given-a-threshold/1283-find-the-smallest-divisor-given-a-threshold.java
@@ -25,7 +25,7 @@ class Solution {
     public int sum(int d , int[] nums){
         int sum = 0;
         for(int val : nums){
-            sum += (int)(Math.ceil(val*1d/ d));
+            sum += (Math.ceil(val*1d/ d));
         }
         return sum;
     }
